requirements.txt
numpy~=1.18.4
pillow~=7.1.2
torch~=1.5.0
torchvision~=0.4.2